+++ src/Trarizon.Library/Collections/Specialized/Memento.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Trarizon.Library.Collections.StackAlloc;
 
-namespace Trarizon.Library.Collections.Generic;
+namespace Trarizon.Library.Collections.Specialized;
 public class Memento<T>
 {
     private T[] _array;